real fix to the tab variable problem: changed the c_types int to atb_type mapping from 16 to 32 bit. alternative would have been using c_longs but on mac and ubuntu an int is 32 bit and the same as a long, which makes the int = atb_int16 mapping confusing.

+++ Capture/world.py
@@ -1,6 +1,6 @@
 import os, sys
 
-from ctypes import  c_int,c_int64,c_bool,c_float
+from ctypes import  c_int,c_bool,c_float
 import numpy as np
 from glob import glob
 import cv2
@@ -156,12 +156,12 @@ def world(src, size, g_pool):
             bar.camera_state[control["name"]]=c_bool(control["value"])
             bar.add_var("Camera/"+control["name"],bar.camera_ctl[control["name"]])
         elif control["type"]=="(int)":
-            bar.camera_ctl[control["name"]]=c_int64(control["value"])
-            bar.camera_state[control["name"]]=c_int64(control["value"])
+            bar.camera_ctl[control["name"]]=c_int(control["value"])
+            bar.camera_state[control["name"]]=c_int(control["value"])
             bar.add_var("Camera/"+control["name"],bar.camera_ctl[control["name"]],max=control["max"],min=control["min"],step=control["step"])
         elif control["type"]=="(menu)":
-            bar.camera_ctl[control["name"]]=c_int64(control["value"])
-            bar.camera_state[control["name"]]=c_int64(control["value"])
+            bar.camera_ctl[control["name"]]=c_int(control["value"])
+            bar.camera_state[control["name"]]=c_int(control["value"])
             bar.add_var("Camera/"+control["name"],bar.camera_ctl[control["name"]],max=control["max"],min=control["min"],step=1)
         else:
             pass
